pyproject.toml
[tool.poetry]
name = "cafe"
version = "0.1.0"
description = ""
authors = ["Artyom_Vorozhbyanov <artyom.vorozhbyanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kivy = "^2.1.0"
kivymd = "^1.1.1"
django = "^4.1.7"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
twilio = "^7.16.4"
drf-nested-routers = "^0.93.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
